Adds a base activity to use the drawer
