+++ app/src/main/res/layout/activity_dh_homepage.xml
@@ -1,101 +1,49 @@
 <?xml version="1.0" encoding="utf-8"?>
-<android.support.v4.widget.DrawerLayout
+<LinearLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
-    android:id="@+id/dh_homepage_drawer_layout"
+    android:id="@+id/dh_homepage_content_frame"
+    android:orientation="vertical"
     android:layout_width="match_parent"
     android:layout_height="match_parent">
 
-    <!-- As the main content view, the view below consumes the entire
-         space available using match_parent in both dimensions. -->
-    <LinearLayout
-        android:id="@+id/dh_homepage_content_frame"
-        android:orientation="vertical"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
+<Button
+    android:id="@+id/dh_homepage_btn_newdecision"
+    android:text="@string/dh_homepage_btn_newdecisiontext"
+    android:layout_width="match_parent"
+    android:layout_height="wrap_content"
+    android:textSize="20sp"
+    android:layout_marginTop="30dp"
+    android:onClick="dh_homepage_btn_newdecisionOnClick"/>
 
-        <android.support.v7.widget.Toolbar
-            android:id="@+id/dh_homepage_ab"
-            android:layout_width="match_parent"
-            android:layout_height="40dp"
-            android:background="@color/colorPrimary">
-            <TextView
-                android:id="@+id/dh_homepage_tv_titlebar"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:text="@string/dh_homepage_titlebar"
-                android:textColor="#ffffff"
-                android:textSize = "26sp"
-                android:gravity = "center" />
-        </android.support.v7.widget.Toolbar>
+<TextView
+    android:id="@+id/dh_homepage_tv_mydecisions"
+    android:text="@string/dh_homepage_tv_mydecisionstext"
+    android:layout_width="match_parent"
+    android:layout_height="wrap_content"
+    android:layout_marginTop="30dp"
+    android:textSize="20sp" />
+
 
 
     <Button
-        android:id="@+id/dh_homepage_btn_newdecision"
-        android:text="@string/dh_homepage_btn_newdecisiontext"
+        android:text="Decide which job to take"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:textSize="20sp"
-        android:layout_marginTop="30dp"
-        android:onClick="dh_homepage_btn_newdecisionOnClick"/>
+        android:textSize="16sp"
+        android:onClick="dh_homepage_btn_jobdecisionOnClick"/>
+
+
+
 
     <TextView
-        android:id="@+id/dh_homepage_tv_mydecisions"
-        android:text="@string/dh_homepage_tv_mydecisionstext"
+        android:id="@+id/dh_homepage_tv_sharedwithme"
+        android:text="@string/dh_homepage_tv_sharedwithme"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_marginTop="30dp"
         android:textSize="20sp" />
 
 
-
-        <Button
-            android:text="Decide which job to take"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:textSize="16sp"
-            android:onClick="dh_homepage_btn_jobdecisionOnClick"/>
-
-
-
-
-        <TextView
-            android:id="@+id/dh_homepage_tv_sharedwithme"
-            android:text="@string/dh_homepage_tv_sharedwithme"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-			android:layout_marginTop="30dp"
-            android:textSize="20sp" />
-
-
-
-    </LinearLayout>
-
-
-    <!-- The navigation drawer -->
-    <RelativeLayout
-        android:id="@+id/dh_homepage_drawerPane"
-        android:layout_width="220dp"
-        android:layout_height="match_parent"
-        android:layout_gravity="start">
-
-
-    <!-- android:layout_gravity="start" tells DrawerLayout to treat
-         this as a sliding drawer on the left side for left-to-right
-         languages and on the right side for right-to-left languages.
-         The drawer is given a fixed width in dp and extends the full height of
-         the container. A solid background is used for contrast
-         with the content view. -->
-    <ListView
-        android:id="@+id/dh_homepage_navList"
-        android:layout_width="240dp"
-        android:layout_height="match_parent"
-        android:layout_gravity="start"
-        android:choiceMode="singleChoice"
-        android:divider="@android:color/transparent"
-        android:dividerHeight="0dp"
-        android:background="@color/colorWhite"/>
-
-    </RelativeLayout>
-</android.support.v4.widget.DrawerLayout>
+</LinearLayout>
